Moved code to utility from views ccx
